chromium: rust: Use adler instead of adler2
Upstream Rust replaced adler with adler2, for older versions of Rust we still need
to tell GN that we have the older lib when it tries to copy the Rust sysroot
into the bulid directory.

It looks that in rust-1.85.0, `adler` has not yet been replaced with `adler2`.
Therefore, the condition should likely be adjusted.

This patch will be dropped once Rust >= 1.86.0 is available, which
replaced adler with adler2.

Note that currently meta-lts-mixins' `scarthgap/rust` branch [0]
provides Rust version `1.85.1`.

[0] https://git.yoctoproject.org/meta-lts-mixins/log/?h=scarthgap/rust

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@collabora.com>